depth-anything/requirements.txt
# Depth Anything V2 Dependencies
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
PyYAML>=6.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.35.0
accelerate>=0.25.0

# Point cloud generation
open3d>=0.17.0

